requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.4.0
    # via httpx
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
chardet==5.2.0
    # via trimesh
colorlog==6.8.2
    # via trimesh
embreex==2.17.7.post4
    # via trimesh
exceptiongroup==1.2.1
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via trimesh
idna==3.7
    # via
    #   anyio
    #   httpx
jsonschema==4.22.0
    # via trimesh
jsonschema-specifications==2023.12.1
    # via jsonschema
libigl==2.5.1
    # via -r requirements.in
lxml==5.2.2
    # via trimesh
mapbox-earcut==1.0.1
    # via trimesh
networkx==3.3
    # via trimesh
numpy==1.26.4
    # via
    #   -r requirements.in
    #   embreex
    #   libigl
    #   mapbox-earcut
    #   pycollada
    #   scipy
    #   shapely
    #   trimesh
    #   vhacdx
    #   xatlas
pillow==10.3.0
    # via trimesh
pycollada==0.8
    # via trimesh
python-dateutil==2.9.0.post0
    # via pycollada
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
rtree==1.2.0
    # via trimesh
scipy==1.13.1
    # via
    #   -r requirements.in
    #   libigl
    #   trimesh
shapely==2.0.4
    # via trimesh
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
svg-path==6.3
    # via trimesh
trimesh[easy]==4.4.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via anyio
vhacdx==0.0.8
    # via trimesh
xatlas==0.0.9
    # via trimesh
xxhash==3.4.1
    # via trimesh

# The following packages are considered to be unsafe in a requirements file:
# setuptools
